# api/public/docs/openapi.yml
openapi: 3.1.0
info:
  title: 'Constellation APIs'
  license:
    name: 'GNU AFFERO GENERAL PUBLIC LICENSE'
    url: 'https://www.gnu.org/licenses/agpl-3.0.en.html'
  version: '2.0'
servers:
  -
    url: 'http://localhost:8080/api'
    description: 'The local environment.'
paths:
...
  /v1/teams:
    get:
      tags:
        - Team
      summary: 'Get a listing of the teams'
      description: 'Get all teams.'
      operationId: getTeams
      parameters:
        -
          $ref: '#/components/parameters/base-filter-per-page'
        -
          $ref: '#/components/parameters/base-filter-page'
        -
          $ref: '#/components/parameters/base-filter-sort'
        -
          $ref: '#/components/parameters/base-filter-q'
        -
          name: 'filter[id]'
          in: query
          description: 'Filter by team id.'
          schema:
            type: integer
      responses:
        '200':
          description: 'Teams list'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-team' } }
                  message: { type: string }
                  total: { type: integer }
                type: object
    post:
      tags:
        - Team
      summary: 'Store an team'
      description: 'Store an team.'
      operationId: storeTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-team'
      responses:
        '200':
          description: 'Created team data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-team' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
  '/v1/teams/{id}':
    get:
      tags:
        - Team
      summary: 'Display the specified team'
      description: 'Get an team.'
      operationId: showTeam
      parameters:
        -
          name: id
          in: path
          description: 'id of the team'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Team detail'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/resource-team' } }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    put:
      tags:
        - Team
      summary: 'Update a team'
      description: 'Update a team.'
      operationId: updateTeam
      parameters:
        -
          name: id
          in: path
          description: 'id of the team'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request-create-team'
      responses:
        '200':
          description: 'Updated team data.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/resource-team' }
                  message: { type: string }
                type: object
        '422':
          description: 'Invalid inputs.'
          content:
            application/json:
              schema:
                properties:
                  errors: { properties: { '{name-of-invalid-field}': { type: array, items: { description: 'Error message', type: string } } }, type: object }
                  message: { type: string }
                type: object
        '404':
          description: 'Resource not found'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                  exception: { type: string }
                  file: { type: string }
                  line: { type: integer }
                  trace: { type: array, items: { properties: { file: { type: string }, line: { type: integer }, function: { type: string }, class: { type: string }, type: { type: string } }, type: object } }
                type: object
    delete:
      tags:
        - Team
      summary: 'Delete an team'
      description: 'Remove the specified Team from storage.'
      operationId: deleteTeam
      parameters:
        -
          name: id
          in: path
          description: 'id of the team'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Team deleted.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
        '404':
          description: 'Team not found.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
...
components:
  schemas:
...
    request-create-team:
      title: 'Create team'
      description: 'Create team request'
      required:
        - name
        - manager
      properties:
        name:
          description: Name
          type: string
        manager:
          description: 'Manager name'
          type: string
      type: object
...
    resource-team:
      title: Team
      description: 'Team resource'
      properties:
        id:
          description: id
          type: integer
        name:
          description: name
          type: string
        manager:
          description: manager
          type: string
        created_at:
          description: created_at
          type: string
          format: date-time
        updated_at:
          description: updated_at
          type: string
          format: date-time
      type: object
...
  parameters:
    base-filter-per-page:
      name: perPage
      in: query
      description: 'Number of items displayed by page.'
      schema:
        type: integer
        default: 100
        enum:
          - 5
          - 10
          - 25
          - 100
          - 1000
    base-filter-page:
      name: page
      in: query
      description: 'Current page.'
      schema:
        type: integer
    base-filter-sort:
      name: sort
      in: query
      description: 'Field used to sort the result.'
      schema:
        type: string
      examples:
        string:
          summary: '-id'
          value: '-id'
    base-filter-q:
      name: q
      in: query
      description: 'Query string'
      required: false
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Basic Auth'
      scheme: bearer
security:
  -
    bearerAuth: []